-now compiled with devkitpro ppc r29-1 and libogc 1.8.17 -updated libcustomfat and libvorbis thanks to GreyWolf -updated libturbojpeg thanks to Fix94 -fixed cheat menu so cheat names with apostrophe's display properly

+++ trunk/source/channel/MD5.h
@@ -88,7 +88,7 @@ extern "C"
 typedef struct
   {
   unsigned int len;
-  unsigned int ABCD[4];
+  unsigned long int ABCD[4];
   int      b_used;
   unsigned char    block[64];
   } auth_md5Ctx;
